requirements.txt
flask
flask-cors
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
pytest
pytest-playwright